Fix: bug in actor logit's numerical precision when bf16 is on (#634)

+++ openrlhf/models/actor.py
@@ -211,7 +211,7 @@ class Actor(nn.Module):
             assert return_output
             return output
 
-        log_probs = log_probs_from_logits(output["logits"][:, :-1, :], sequences[:, 1:])
+        log_probs = log_probs_from_logits(output["logits"][:, :-1, :].to(torch.float32), sequences[:, 1:])
 
         if not self.packing_samples:
             action_log_probs = log_probs[:, -num_actions:]
